requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
docusign-esign==4.0.0
exceptiongroup==1.1.1
Flask==3.0.3
gunicorn==23.0.0
Jinja2==3.1.4
num2words==0.5.13
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dotenv==1.0.0
pywin32==306
requests==2.31.0
tenacity==9.0.0

